feat(candles): add @typedtrader/candles fixtures package (#1125)
