+++ packages/candles/json/downtrend-msft.json
