+++ packages/candles/json/downtrend.json
