+++ packages/candles/json/uptrend-slv.json
