+++ packages/candles/json/uptrend.json
