+++ packages/candles/tsconfig.build.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.build.json",
+  "compilerOptions": {
+    "rootDir": "${configDir}"
+  },
+  "include": ["${configDir}/src", "${configDir}/json"]
+}
