+++ packages/trading-signals-docs/package.json
@@ -3,6 +3,7 @@
     "@highcharts/react": "^5.0.1",
     "@tailwindcss/postcss": "^4.3.0",
     "@testing-library/dom": "^10.4.1",
+    "@typedtrader/candles": "0.1.0",
     "@typedtrader/exchange": "0.2.5",
     "highcharts": "^12.6.0",
     "next": "^16.2.6",
